GlassDigSite updated

+++ SonOfRobin.Core/Game/Animation/AnimFrame.cs
@@ -17,7 +17,7 @@ namespace SonOfRobin
         public readonly Vector2 gfxOffset;
         public readonly Vector2 colOffset;
         public readonly byte layer;
-        public readonly byte duration;
+        public readonly int duration;
         public readonly float scale;
         public readonly bool ignoreWhenCalculatingMaxSize;
         public readonly Texture2D texture;
@@ -25,7 +25,7 @@ namespace SonOfRobin
         public readonly Rectangle textureRect;
         public readonly Vector2 rotationOrigin;
 
-        public static AnimFrame GetFrame(string atlasName, int atlasX, int atlasY, int width, int height, byte layer, byte duration, bool crop = false, float scale = 1f, float depthPercent = 0.25f, int padding = 1, bool ignoreWhenCalculatingMaxSize = false)
+        public static AnimFrame GetFrame(string atlasName, int atlasX, int atlasY, int width, int height, byte layer, int duration, bool crop = false, float scale = 1f, float depthPercent = 0.25f, int padding = 1, bool ignoreWhenCalculatingMaxSize = false)
         {
             // some frames are duplicated and can be reused (this can be verified by checking ID)
 
@@ -35,12 +35,12 @@ namespace SonOfRobin
             else return new AnimFrame(atlasName: atlasName, atlasX: atlasX, atlasY: atlasY, width: width, height: height, layer: layer, duration: duration, crop: crop, scale: scale, depthPercent: depthPercent, padding: padding, ignoreWhenCalculatingMaxSize: ignoreWhenCalculatingMaxSize);
         }
 
-        private static string GetID(string atlasName, int atlasX, int atlasY, int width, int height, byte layer, byte duration, bool crop, float scale, float depthPercent)
+        private static string GetID(string atlasName, int atlasX, int atlasY, int width, int height, byte layer, int duration, bool crop, float scale, float depthPercent)
         {
             return $"{atlasName}_{atlasX},{atlasY}_{width}x{height}_{layer}_{duration}_{crop}_{scale}_{depthPercent}";
         }
 
-        private AnimFrame(string atlasName, int atlasX, int atlasY, int width, int height, byte layer, byte duration, bool crop, float scale, float depthPercent, int padding, bool ignoreWhenCalculatingMaxSize)
+        private AnimFrame(string atlasName, int atlasX, int atlasY, int width, int height, byte layer, int duration, bool crop, float scale, float depthPercent, int padding, bool ignoreWhenCalculatingMaxSize)
         {
             // should not be invoked from other classes directly
